+++ src/lib.rs
@@ -41,7 +41,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! typed-index-collections = "3.4.0"
+//! typed-index-collections = "3.5.0"
 //! ```
 //!
 //! This crate depends on the standard library by default that is useful
@@ -51,7 +51,7 @@
 //!
 //! ```toml
 //! [dependencies.typed-index-collections]
-//! version = "3.4.0"
+//! version = "3.5.0"
 //! default-features = false
 //! features = ["alloc"]
 //! ```
@@ -63,7 +63,7 @@
 //! ```toml
 //! [dependencies]
 //! derive_more = "0.99"
-//! typed-index-collections = "3.4.0"
+//! typed-index-collections = "3.5.0"
 //! ```
 //!
 //! # Examples