+++ CHANGELOG.md
@@ -6,12 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.5.0] - 2026-01-18
 ### Added
 - `TiVec::extend_from_within_corrected` method, a corrected variant of
   `TiVec::extend_from_within` method.
 
 ### Changed
 - `TiVec::extend_from_within` is marked deprecated.
+- Minore documentation improvements.
 
 ## [3.4.0] - 2025-12-05
 ### Added
@@ -177,7 +180,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `TiSlice` API compatibility tests.
 - Crate API documentation with examples.
 
-[Unreleased]: https://github.com/zheland/typed-index-collections/compare/v3.4.0...HEAD
+[Unreleased]: https://github.com/zheland/typed-index-collections/compare/v3.5.0...HEAD
+[3.5.0]: https://github.com/zheland/typed-index-collections/compare/v3.4.0...v3.5.0
 [3.4.0]: https://github.com/zheland/typed-index-collections/compare/v3.3.0...v3.4.0
 [3.3.0]: https://github.com/zheland/typed-index-collections/compare/v3.2.3...v3.3.0
 [3.2.3]: https://github.com/zheland/typed-index-collections/compare/v3.2.2...v3.2.3